pyproject.toml
[project]
authors = [{name = "Griswald Brooks", email = "griswald.brooks@gmail.com"}]
dependencies = [
    "numpy",
    "onnxruntime",
    "Pillow",
]
name = "cat-detector-scripts"
requires-python = ">= 3.11"
version = "0.1.0"

[project.optional-dependencies]
clip = [
    "torch",
    "clip @ git+https://github.com/openai/CLIP.git",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.pypi-dependencies]
cat-detector-scripts = { path = ".", editable = true }

[tool.pixi.environments]
default = { solve-group = "default" }
clip = { features = ["clip"], solve-group = "default" }
docs = { features = ["docs"], no-default-feature = true, solve-group = "docs" }

[tool.pixi.tasks]
test-clip = "python scripts/test_clip_detection.py"
generate-image-embeddings = """python scripts/generate_image_embeddings.py \
  --classes cat_overhead:prototype_images/cat_overhead \
            empty_room:prototype_images/empty_room_overhead \
            person_overhead:prototype_images/person_overhead \
            litter_box:prototype_images/litter_box_overhead"""

[tool.pixi.feature.clip.tasks]
generate-embeddings = "python scripts/generate_clip_embeddings.py"

[tool.pixi.feature.docs.dependencies]
nodejs = ">=22"

[tool.pixi.feature.docs.tasks]
docs-install = { cmd = "npm ci", cwd = "website" }
docs-build = { cmd = "npm run build", cwd = "website", depends-on = ["docs-install"] }
docs-serve = { cmd = "npm run serve", cwd = "website", depends-on = ["docs-install"] }
docs-start = { cmd = "npm run start", cwd = "website", depends-on = ["docs-install"] }
